fixup! ASoC: SOF: Implement snd_sof_bytes_ext_volatile_get kcontrol IO
remove cast as suggested by Marc Herbert

+++ sound/soc/sof/control.c
@@ -390,7 +390,7 @@ int snd_sof_bytes_ext_volatile_get(struct snd_kcontrol *kcontrol, unsigned int _
 	if (cdata->data->size > be->max - sizeof(const struct sof_abi_hdr)) {
 		dev_err_ratelimited(scomp->dev, "error: user data size %d exceeds max size %zu.\n",
 				    cdata->data->size,
-				    (size_t)be->max - sizeof(const struct sof_abi_hdr));
+				    be->max - sizeof(const struct sof_abi_hdr));
 		ret = -EINVAL;
 		goto out;
 	}
